API: move the functions used to implement the API to a separate file
