+++ makefile
@@ -26,7 +26,7 @@ LIBS=-lelf -lpthread
 HEADERS=*.h makefile
 INCLUDES=-I/usr/include/libelf
 SOURCES= dispatcher.S common.c dbm.c traces.c syscalls.c dispatcher.c signals.c util.S
-SOURCES+=api/helpers.c api/plugin_support.c api/branch_decoder_support.c api/load_store.c
+SOURCES+=api/helpers.c api/plugin_support.c api/branch_decoder_support.c api/load_store.c api/internal.c
 SOURCES+=elf/elf_loader.o
 
 ARCH=$(shell $(CC) -dumpmachine | awk -F '-' '{print $$1}')
